Add icons for pause and a CSS-animated filling pie chart
Pie chart filling animation works by using a thick stroke around a
circle, which appears as the filled area of the pie chart. Using a
dashed stroke with a custom dash width gives the appearance of a pie
chart filled to various levels. An animation fills the pie chart from
empty to full